add support for script-path
This commit also adds better test coverage, mainly for defaulting rules.

Note 1: we have to specify `Jenkinsfile` as a default value. The Jenkins UI says that 'Jenkinsfile' is the default value when the field is left blank, but perhaps that only applies to the UI. When a blank is used in the xml config, the plugin looks for a file named '' (i.e., the empty string.)

Note 2: I'm not sure if the factory's 'owner' element is necessary, but I've discovered that it's needed for other things (mainly branch sources, which will be added in a few commits).