Generate an 'id' for the github branch source element
Jenkins used to handle this fine, even though it apparently violated the contract
(based on some of Stephen Conolly's comments in the jira linked to here).
Jenkins doesn't handle this anymore, with a
"Could not determine exact tip revision of ..." error message.

In some cases the id is supposed to reflect any config options,
but I am not sure our use case is one of those.
I'm assuming it is, so the id we'll generate here should be stable but should change
if any of the traits change.